pyproject.toml
[tool.poetry]
name = "internship"
version = "0.1.0"
description = ""
authors = ["p0lygun <solankivibhakar82@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
markdown = "^3.5"
django-filter = "^23.3"
djangorestframework-simplejwt = "^5.3.0"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
